Move detailsPanels to common and use in investigation card view (#1057)
