+++ packages/datagateway-common/src/detailsPanels/isis/instrumentDetailsPanel.component.test.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { createMount } from '@material-ui/core/test-utils';
 import InstrumentDetailsPanel from './instrumentDetailsPanel.component';
-import { Instrument, useInstrumentDetails } from 'datagateway-common';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { ReactWrapper } from 'enzyme';
+import { Instrument } from '../../app.types';
+import { useInstrumentDetails } from '../../api/instruments';
 
 jest.mock('datagateway-common', () => {
   const originalModule = jest.requireActual('datagateway-common');
