+++ packages/datagateway-common/src/detailsPanels/isis/investigationDetailsPanel.component.test.tsx
@@ -1,13 +1,10 @@
 import React from 'react';
 import { createMount } from '@material-ui/core/test-utils';
-import VisitDetailsPanel from './visitDetailsPanel.component';
-import {
-  Investigation,
-  useInvestigationDetails,
-  useInvestigationSize,
-} from 'datagateway-common';
+import InvestigationDetailsPanel from './investigationDetailsPanel.component';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { ReactWrapper } from 'enzyme';
+import { Investigation } from '../../app.types';
+import { useInvestigationDetails } from '../../api/investigations';
 
 jest.mock('datagateway-common', () => {
   const originalModule = jest.requireActual('datagateway-common');
@@ -16,19 +13,19 @@ jest.mock('datagateway-common', () => {
     __esModule: true,
     ...originalModule,
     useInvestigationDetails: jest.fn(),
-    useInvestigationSize: jest.fn(),
   };
 });
 
-describe('Visit details panel component', () => {
+describe('Investigation details panel component', () => {
   let mount;
   let rowData: Investigation;
   const detailsPanelResize = jest.fn();
+  const viewDatasets = jest.fn();
 
   const createWrapper = (): ReactWrapper => {
     return mount(
       <QueryClientProvider client={new QueryClient()}>
-        <VisitDetailsPanel
+        <InvestigationDetailsPanel
           rowData={rowData}
           detailsPanelResize={detailsPanelResize}
         />
@@ -55,16 +52,21 @@ describe('Visit details panel component', () => {
           },
         },
       ],
+      studyInvestigations: [
+        {
+          id: 11,
+          study: {
+            id: 12,
+            pid: 'study pid',
+          },
+        },
+      ],
       startDate: '2019-06-10',
       endDate: '2019-06-11',
     };
 
     (useInvestigationDetails as jest.Mock).mockReturnValue({
       data: rowData,
-    });
-    (useInvestigationSize as jest.Mock).mockReturnValue({
-      data: 1,
-      refetch: jest.fn(),
     });
   });
 
@@ -75,7 +77,7 @@ describe('Visit details panel component', () => {
 
   it('renders correctly', () => {
     const wrapper = createWrapper();
-    expect(wrapper.find('VisitDetailsPanel').props()).toMatchSnapshot();
+    expect(wrapper.find('InvestigationDetailsPanel').props()).toMatchSnapshot();
   });
 
   it('renders user, sample and publication tabs when present in the data', () => {
@@ -114,7 +116,7 @@ describe('Visit details panel component', () => {
     ];
 
     const wrapper = createWrapper();
-    expect(wrapper.find('VisitDetailsPanel').props()).toMatchSnapshot();
+    expect(wrapper.find('InvestigationDetailsPanel').props()).toMatchSnapshot();
   });
 
   it('checks if multiple samples result in change of title to plural version', () => {
@@ -130,7 +132,7 @@ describe('Visit details panel component', () => {
     ];
 
     const wrapper = createWrapper();
-    expect(wrapper.find('VisitDetailsPanel').props()).toMatchSnapshot();
+    expect(wrapper.find('InvestigationDetailsPanel').props()).toMatchSnapshot();
   });
 
   it('checks if multiple publications result in change of title to plural version', () => {
@@ -146,7 +148,7 @@ describe('Visit details panel component', () => {
     ];
 
     const wrapper = createWrapper();
-    expect(wrapper.find('VisitDetailsPanel').props()).toMatchSnapshot();
+    expect(wrapper.find('InvestigationDetailsPanel').props()).toMatchSnapshot();
   });
 
   it('renders publication tab and text "No Publications" when no data is prsent', () => {
@@ -154,7 +156,7 @@ describe('Visit details panel component', () => {
     const wrapper = createWrapper();
     expect(
       wrapper
-        .find('[data-testid="visit-details-panel-no-publications"]')
+        .find('[data-testid="investigation-details-panel-no-publications"]')
         .exists()
     ).toBeTruthy();
   });
@@ -163,7 +165,9 @@ describe('Visit details panel component', () => {
     rowData.samples = [];
     const wrapper = createWrapper();
     expect(
-      wrapper.find('[data-testid="visit-details-panel-no-samples"]').exists()
+      wrapper
+        .find('[data-testid="investigation-details-panel-no-samples"]')
+        .exists()
     ).toBeTruthy();
   });
 
@@ -171,31 +175,15 @@ describe('Visit details panel component', () => {
     rowData.investigationUsers = [];
     const wrapper = createWrapper();
     expect(
-      wrapper.find('[data-testid="visit-details-panel-no-name"]').exists()
+      wrapper
+        .find('[data-testid="investigation-details-panel-no-name"]')
+        .exists()
     ).toBeTruthy();
   });
 
-  it('calls useInvestigationDetails and useInvestigationSize hooks on load', () => {
+  it('calls useInvestigationDetails hook on load', () => {
     createWrapper();
     expect(useInvestigationDetails).toHaveBeenCalledWith(rowData.id);
-    expect(useInvestigationSize).toHaveBeenCalledWith(rowData.id);
-  });
-
-  it('renders calculate size button when size has not been calculated', () => {
-    (useInvestigationSize as jest.Mock).mockReturnValueOnce({});
-    const wrapper = createWrapper();
-    expect(wrapper.find('#calculate-size-btn').exists()).toBeTruthy();
-  });
-
-  it('calculates size when button is clicked', () => {
-    const fetchSize = jest.fn();
-    (useInvestigationSize as jest.Mock).mockReturnValueOnce({
-      refetch: fetchSize,
-    });
-
-    const wrapper = createWrapper();
-    wrapper.find('#calculate-size-btn').hostNodes().simulate('click');
-    expect(fetchSize).toHaveBeenCalled();
   });
 
   it('calls detailsPanelResize on load and when tabs are switched between', () => {
@@ -210,7 +198,10 @@ describe('Visit details panel component', () => {
 
     expect(detailsPanelResize).toHaveBeenCalledTimes(1);
 
-    wrapper.find('#visit-publications-tab').hostNodes().simulate('click');
+    wrapper
+      .find('#investigation-publications-tab')
+      .hostNodes()
+      .simulate('click');
 
     expect(detailsPanelResize).toHaveBeenCalledTimes(2);
   });
@@ -225,18 +216,43 @@ describe('Visit details panel component', () => {
 
     const wrapper = mount(
       <QueryClientProvider client={new QueryClient()}>
-        <VisitDetailsPanel rowData={rowData} />
+        <InvestigationDetailsPanel rowData={rowData} />
       </QueryClientProvider>
     );
 
     expect(detailsPanelResize).toHaveBeenCalledTimes(0);
 
-    wrapper.find('#visit-publications-tab').hostNodes().simulate('click');
+    wrapper
+      .find('#investigation-publications-tab')
+      .hostNodes()
+      .simulate('click');
 
     expect(detailsPanelResize).toHaveBeenCalledTimes(0);
   });
 
-  it('gracefully handles InvestigationUsers without Users', () => {
+  it('displays DOI and renders the expected Link ', () => {
+    const wrapper = createWrapper();
+    expect(
+      wrapper
+        .find('[data-testid="investigation-details-panel-doi-link"]')
+        .first()
+        .text()
+    ).toEqual('doi 1');
+    expect(
+      wrapper
+        .find('[data-testid="investigation-details-panel-doi-link"]')
+        .first()
+        .prop('href')
+    ).toEqual('https://doi.org/doi 1');
+  });
+
+  it('gracefully handles StudyInvestigations without Studies and InvestigationUsers without Users', () => {
+    rowData.studyInvestigations = [
+      {
+        id: 11,
+      },
+    ];
+
     rowData.investigationUsers = [
       {
         id: 4,
@@ -249,11 +265,27 @@ describe('Visit details panel component', () => {
     });
 
     const wrapper = createWrapper();
-    expect(wrapper.find('VisitDetailsPanel').props()).toMatchSnapshot();
+    expect(wrapper.find('InvestigationDetailsPanel').props()).toMatchSnapshot();
+  });
+
+  it('calls dataset view if view datasets tab clicked', () => {
+    const wrapper = mount(
+      <QueryClientProvider client={new QueryClient()}>
+        <InvestigationDetailsPanel
+          rowData={rowData}
+          detailsPanelResize={detailsPanelResize}
+          viewDatasets={viewDatasets}
+        />
+      </QueryClientProvider>
+    );
+
+    expect(viewDatasets).not.toHaveBeenCalled();
+    wrapper.find('#investigation-datasets-tab').hostNodes().simulate('click');
+    expect(viewDatasets).toHaveBeenCalled();
   });
 
   it('Shows "No <field> provided" incase of a null field', () => {
-    const { summary, startDate, endDate, ...amendedRowData } = rowData;
+    const { summary, doi, startDate, endDate, ...amendedRowData } = rowData;
 
     (useInvestigationDetails as jest.Mock).mockReturnValueOnce({
       data: amendedRowData,
@@ -261,7 +293,7 @@ describe('Visit details panel component', () => {
 
     const wrapper = mount(
       <QueryClientProvider client={new QueryClient()}>
-        <VisitDetailsPanel
+        <InvestigationDetailsPanel
           rowData={amendedRowData}
           detailsPanelResize={detailsPanelResize}
         />
@@ -269,13 +301,16 @@ describe('Visit details panel component', () => {
     );
 
     expect(wrapper.html()).toContain(
+      '<b>investigations.details.summary not provided</b>'
+    );
+    expect(wrapper.html()).toContain(
+      '<b>investigations.details.doi not provided</b>'
+    );
+    expect(wrapper.html()).toContain(
       '<b>investigations.details.start_date not provided</b>'
     );
     expect(wrapper.html()).toContain(
       '<b>investigations.details.end_date not provided</b>'
-    );
-    expect(wrapper.html()).toContain(
-      '<b>investigations.details.summary not provided</b>'
     );
   });
 });
